% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateCounts.R
\name{gateCountAdjustment}
\alias{gateCountAdjustment}
\title{Calculates Cumulative Gate Counts}
\usage{
gateCountAdjustment(
  rawGateCounts,
  gateType = "Unidirectional",
  gatecounterMaxValue = 999999
)
}
\arguments{
\item{rawGateCounts}{A numeric vector of length corresponding to days or
are present.}

\item{gateType}{A character string with options "Unidirectional" or}

\item{gatecounterMaxValue}{A numeric value greater than 0 indicating the
gate counter max value, before it is reset. The default value is
999,999.}
}
\value{
Returns an S3 object of class InfCriteria with results.
\itemize{
  \item adjustedCountSum - Sum of daily gate counts for the period,
        adjusted for issues mentioned under details. If gateType was
        "Bidirectional", the final resulting number would be divided
        by two.
  \item unadjustedDailyCounts - A vector of daily counts, not adjusted
        for gate type. If "Bidirectional" gate type, then would need to
        sum this vector and divide by two at the end.
  \item gateType - Gate type for which counts are provided by the user.
}
}
\description{
A function that calculates cumulative gate counts, provided a numeric
vector or a tibble containing values of raw daily gate counts. The
function adjusts for several factors outlined under details.
}
\details{
The function requires directionality of the gates for which
   the daily counts are provided (called gate type). If unidirectional
   no adjustments are done. If bidirectional, the cumulative sum at
   the end will be divided by two. Testing has shown that this method
   of dividing at the end will help reduce issues with counts
   that result from division of daily counts by two and rounding up or
   down. Further, negative counts can result during calculation if
   the counter has reset or if a lower value has been entered compared
   to previous day. This function attempts to correct for both scenarios.
   In the case of counter reset, the 'gatecounterMaxValue' argument
   provided by the user will be used to calculate the corrected value.
   In the case of a lower value being entered compared to previous day,
   the negative value will be replaced by NA. For next day, the count is
   calculated by the numeric value previously reported prior to NA, that
   must be higher than current value. The function also ensure counts
   for empty cells (when the count was forgotten to be reported) are
   accounted for. All scenarios are explained with images on the tutorial.
}
\examples{
# Example 1:
randomCounts <- c(sort(rpois(n = 50, lambda = 100)),
                  sort(rpois(n = 50, lambda = 1000)),
                  sort(rpois(n = 82, lambda = 100000)),
                       200000, # max value
                  sort(rpois(n = 50, lambda = 100)),
                  sort(rpois(n = 50, lambda = 1000)),
                  sort(rpois(n = 83, lambda = 100000)))

rbs1FloorNORTHcount <- gateCountAdjustment(
             rawGateCounts = randomCounts,
             gateType = "Unidirectional",
             gatecounterMaxValue = 200000)
             rbs1FloorNORTHcount$countSum # 70686

}
