% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGateCount.R
\name{shinyGateCount}
\alias{shinyGateCount}
\title{Launch Shiny App For Package gateCounts}
\usage{
shinyGateCount()
}
\value{
No return value but open up a shiny page.
}
\description{
A function that launches the shiny app for gateCounts
package. The shiny app permit to calculate cumulative
gate count with adjustments, where needed. Visualization
can be performed.
}
\examples{
\dontrun{
shinyGateCount()
}

}
\references{
Müller K, Wickham H (2022). _tibble: Simple Data Frames_. R
package version 3.1.8, \href{https://CRAN.R-project.org/package=tibble}{Link}.

Google. (2022, February 14). Cleaning up gate count statistics.
Google Groups. Retrieved September 26, 2022,
\href{https://groups.google.com/a/arl.org/g/arl-assess/c/JQyllZN4gaE}{Link}.
}
\author{
Anjali Silva, \email{a.silva@utoronto.ca}
}
