% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateCountsVisMultipleInstitutes.R
\name{gateCountsVisDaily}
\alias{gateCountsVisDaily}
\title{Visualize Daily Visitor Counts of Multiple Institutes}
\usage{
gateCountsVisDaily(dailyVisitorCount)
}
\arguments{
\item{dailyVisitorCount}{A numeric vector or a tibble, with
number of rows equaling to length of days and columns equaling
to 1 + number of institutes, such that the dimension is: days x 1
+ number of institutes. Here, days is the number of days for
which visitor counts are present.
First column must contain the dates and should contain column name
"dates". Dates must be in the format of date-month-year. The second
column must contain the visitor count for the given date and
should be called "counts1".}
}
\value{
Returns three plots summarizing daily visitor counts by
   week, month and day.
\itemize{
  \item dailyOuput - A bar plot of daily visitor count values
        by month.
  \item dailyOuputLine - A line plot of daily visitor count values
        by month.
  \item dailyOuputLine2 - A line plot of daily visitor count values
        by month, laid over in one view.
  \item dailyOuputLog - A bar plot of daily visitor count values
        by month.
  \item dailyOuputStackedLog - A bar plot of log-transformed daily
  visitor count values by month.
  \item dailyOuputLineLog - A line plot of log-transformed daily
  visitor count values by month.
  \item dailyOuputLine2Log - A line plot of log-transformed daily
  visitor count values by month, laid over in one view.
  \item dailyHeatmap - A heatmap of log-transformed daily
  visitor count values
}
}
\description{
A function that permit to create plots showing daily visitor counts
for multiple insitutes.
Plots include line plots and barplots of month vs daily
counts. Plots with log-transformed counts are also provided
to better visualize trends with outliers.
}
\examples{
set.seed(1234)
# Example 1: Unidirectional gates with daily counts
# Simulate gate count data using Poisson distribution
randomCounts1 <- c(sort(rpois(n = 50, lambda = 10000)),
                  sort(rpois(n = 50, lambda = 1000)),
                  sort(rpois(n = 82, lambda = 100)),
                  sort(rpois(n = 50, lambda = 10)),
                  sort(rpois(n = 50, lambda = 1000)),
                  sort(rpois(n = 50, lambda = 10000)))

# Create a tibble with date information
randomCounts1tibble <- tibble::tibble(
                        dates = seq(lubridate::dmy('01-01-2022'),
                        lubridate::dmy('31-12-2022'),
                        by='1 day')[1:length(randomCounts1)] \%>\%
                        format('\%d-\%m-\%Y'),
                        counts = randomCounts1)

set.seed(2345)
randomCounts2 <- c(sort(rpois(n = 50, lambda = 10)),
                  sort(rpois(n = 50, lambda = 50)),
                  sort(rpois(n = 82, lambda = 100)),
                  sort(rpois(n = 50, lambda = 10)),
                  sort(rpois(n = 50, lambda = 3)),
                  sort(rpois(n = 50, lambda = 10000)))

# Create a tibble with date information
randomCounts2tibble <- tibble::tibble(
                        dates = seq(lubridate::dmy('01-01-2022'),
                        lubridate::dmy('31-12-2022'),
                        by='1 day')[1:length(randomCounts2)] \%>\%
                        format('\%d-\%m-\%Y'),
                        counts2 = randomCounts2)

# combine data
multipleTibble <- randomCounts1tibble \%>\%
  dplyr::left_join(randomCounts2tibble)

# Check max value for gate counter maximum
max(multipleTibble$counts, na.rm = TRUE) # 10,254
max(multipleTibble$counts2, na.rm = TRUE) # 10,197

# Visualize
visPutEx1 <-
   gateCountsVisDaily(dailyVisitorCount = visitorCountsEx1)


}
\author{
Anjali Silva, \email{anjali@alumni.uoguelph.ca}
}
