% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeGateCounts.R
\name{gateCountsVisualize}
\alias{gateCountsVisualize}
\title{Visualize Cumulative Daily Visitor Counts and Raw Gate Counts}
\usage{
gateCountsVisualize(outputDailyCounts)
}
\arguments{
\item{outputDailyCounts}{A numeric vector of length corresponding to days or
are present.}
}
\value{
Returns a dotted plot of cumulative daily visitor counts and
   raw gate count. Important: the counts may not reflect daily counts,
   rather adjusted visitor count as of the day. See README file examples
   for details.
}
\description{
A function that permit to create a line plot showing visitor
counts and raw gate counts by day. Important: the counts may
not reflect daily counts, rather adjusted visitor count as
of the day. See README file examples for details.
}
\examples{
set.seed(1234)
# Example 1: Unidirectional gates with daily counts
randomCounts1 <- c(sort(rpois(n = 50, lambda = 100)),
                  sort(rpois(n = 50, lambda = 1000)),
                  sort(rpois(n = 82, lambda = 100000)),
                  200000, # max value
                  sort(rpois(n = 50, lambda = 100)),
                  sort(rpois(n = 50, lambda = 1000)),
                  sort(rpois(n = 50, lambda = 100000)))

randomCounts1tibble <- tibble::tibble(
                        dates = seq(lubridate::dmy('01-01-2022'),
                        lubridate::dmy('31-12-2022'),
                        by='1 day')[1:length(randomCounts1)] \%>\%
                        format('\%d-\%m-\%Y'),
                        counts = randomCounts1)

# check max value for gate counter maximum
max(randomCounts1tibble$counts, na.rm = TRUE) # 200000

randomCountsSumEx1 <- gateCountSummary(
             rawGateCounts = randomCounts1tibble,
             gateType = "Unidirectional",
             gatecounterMaxValue = 200000,
             printMessages = FALSE)
randomCountsSumEx1$dailyCounts # access daily adjusted counts

# Visualize counts from Example 1
visOne <- gateCountsVisualize(
             outputDailyCounts = randomCountsSumEx1)

# Example 2: Bidirectional gates with NA values
randomCounts4 <- c(sort(rpois(n = 50, lambda = 10000)),
                  sort(rpois(n = 50, lambda = 400000)),
                  sort(rpois(n = 82, lambda = 800000)),
                       999999, # max value
                  sort(rpois(n = 50, lambda = 10000)),
                  sort(rpois(n = 50, lambda = 450000)),
                  sort(rpois(n = 50, lambda = 850000)))

# randomly introduce NA and "Gate broken" entries
randomPositions <- sample(x = c(1:length(randomCounts4)),
                         size = 8, replace = FALSE)
randomCounts4[randomPositions[1:4]] <- NA
randomCounts4[randomPositions[5:8]] <- "Gate broken"

randomCountsSumEx4 <- gateCountCumulative(
             rawGateCounts = randomCounts4,
             gateType = "Bidirectional",
             gatecounterMaxValue = 999999)
randomCountsSumEx4$adjustedCountSum # access cumulative count

# Visualize counts from Example 2
visTwo <- gateCountsVisualize(
             rawGateCounts = as.numeric(randomCounts4),
             gateType = "Bidirectional",
             unadjustedDailyCounts = randomCountsSumEx4$unadjustedDailyCounts)


}
\references{
Müller K, Wickham H (2022). _tibble: Simple Data Frames_. R
package version 3.1.8, \href{https://CRAN.R-project.org/package=tibble}{Link}.

Google. (2022, February 14). Cleaning up gate count statistics.
Google Groups. Retrieved September 26, 2022,
\href{https://groups.google.com/a/arl.org/g/arl-assess/c/JQyllZN4gaE}{Link}.
}
\author{
Anjali Silva, \email{anjali@alumni.uoguelph.ca}
}
