% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeGateCounts.R
\name{gateCountsVisualize}
\alias{gateCountsVisualize}
\title{Visualize Cumulative Daily Visitor Counts and Raw Gate Counts}
\usage{
gateCountsVisualize(
  rawGateCounts,
  gateType = "Unidirectional",
  unadjustedDailyCounts
)
}
\arguments{
\item{rawGateCounts}{A numeric vector of length corresponding to days or
are present.}

\item{gateType}{A character string with options "Unidirectional" or}

\item{unadjustedDailyCounts}{A numeric vector of length corresponding to
days, containing values of corrected daily visitor counts, not adjusted
for gate directionality. This would be the 'unadjustedDailyCounts'
output from gateCounts::gateCountCumulative() function in this package.}
}
\value{
Returns a dotted plot of cumulative daily visitor counts and raw gate count.
}
\description{
A function that permit to create a line plot showing cumulative
visitor counts and raw gate counts.
}
\references{
Müller K, Wickham H (2022). _tibble: Simple Data Frames_. R
package version 3.1.8, \href{https://CRAN.R-project.org/package=tibble}{Link}.

Google. (2022, February 14). Cleaning up gate count statistics.
Google Groups. Retrieved September 26, 2022,
\href{https://groups.google.com/a/arl.org/g/arl-assess/c/JQyllZN4gaE}{Link}.
}
\author{
Anjali Silva, \email{anjali@alumni.uoguelph.ca}
}
